refactor: connect ShadCN `Link` to ShadCN `ButtonLink`
